template-parsing-service/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pydantic
supabase
